Extract emission allocation module
